Split import/export and storage preferences
