+++ app/src/main/res/xml/preferences_import_export.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:search="http://schemas.android.com/apk/com.bytehamster.lib.preferencesearch">
+    <Preference
+            android:key="prefOpmlExport"
+            android:title="@string/opml_export_label"/>
+    <Preference
+            android:key="prefOpmlImport"
+            android:title="@string/opml_import_label"/>
+    <Preference
+            android:key="prefHtmlExport"
+            android:title="@string/html_export_label"/>
+    <Preference
+            android:key="importExport"
+            search:keywords="@string/import_export_search_keywords"
+            android:title="@string/import_export"/>
+</PreferenceScreen>
